+++ zscaler/zia/models/urlcategory.py
@@ -40,9 +40,11 @@ class URLCategory(ZscalerObject):
 
             self.urls = ZscalerCollection.form_list(config["urls"] if "urls" in config else [], str)
 
-            self.regex_patterns_retaining_parent_category = ZscalerCollection.form_list(config["regexPatternsRetainingParentCategory"] if "regexPatternsRetainingParentCategory" in config else [], str)
+            self.regex_patterns_retaining_parent_category = ZscalerCollection.form_list
+            (config["regexPatternsRetainingParentCategory"] if "regexPatternsRetainingParentCategory" in config else [], str)
 
-            self.regex_patterns = ZscalerCollection.form_list(config["regexPatterns"] if "regexPatterns" in config else [], str)
+            self.regex_patterns = ZscalerCollection.form_list
+            (config["regexPatterns"]if "regexPatterns" in config else [], str)
 
             self.db_categorized_urls = ZscalerCollection.form_list(
                 config["dbCategorizedUrls"] if "dbCategorizedUrls" in config else [], str
@@ -73,9 +75,7 @@ class URLCategory(ZscalerObject):
             self.description = config["description"] if "description" in config else None
             self.type = config["type"] if "type" in config else None
             self.url_type = config["urlType"] if "urlType" in config else None
-            
-            # Handle nested dictionary for urlKeywordCounts
-            # Handle nested dictionary for urlKeywordCounts
+
             self.url_keyword_counts = (
                 {
                     "totalUrlCount": (